pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cluster-service"
version = "0.1.0"
description = "Cluster Service API — FastAPI clean architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
    "ansi2html>=1.9.2",
    "kubernetes>=31.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.uv]
# uv creates .venv inside the project directory automatically.
# Run: uv sync --group dev
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
